refactor(frontend): reduce DocumentEditor useEffects from 10 to 2
- Absorb read-only content sync into the document machine via pushContentToEditor
  action, making the machine the sole driver of content-to-editor pushes
- Merge _onRootChildrenTypeChange, suppressChangeRef exposure, onEditorReady,
  and imperative handler registration into a single bridge effect
- Extract block-border debug toggle, block highlight, and click-to-edit into
  standalone hooks (useBlockBorderDebug, useBlockHighlight, useReadOnlyClickToEdit)
- Remove document-level Enter-to-edit shortcut; replace with title→summary→editor
  Enter focus chain in EditableDocumentHeader
